day 9 part 1

+++ Day9/Day9.h
@@ -7,9 +7,9 @@ using namespace std;
 
 class Day9 {
 public:
-    static int Part1();
+    static long Part1();
 
-    static int Part2();
+    static long Part2();
 };
 
 #endif //DAY9_H
