+++ Day9/Part2.cpp
@@ -1,6 +1,6 @@
 #include "Day9.h"
 
-int Day9::Part2() {
+long Day9::Part2() {
     const auto lines = Helpers::readFile(9, true);
 
     return 0;
